Testing commission rates with faux inrate contract: writing more tests and renaming files
